Installing NodeJs wit deps role
Refactored the dependencies role to install NodeJs as per owner
documentation and added molecule tests for this role. Addtionally fixed
python-jenkins installation which is a pip package, this involved moving
tasks from the docker role (epel repo and python-pip installs) to this
role so these are triggered earlier in the workflow.